Video course support
Created ability to use a URL argument to show and hide fields (though at the moment they are being removed).  The database also now has an additional field to store what course type the trainee has done.  This as yet is not constrained and is just a free text field - but is fed through from URL parameter through the js scripts